+++ 2-RTB/poc/src/app/globals.css
@@ -27,7 +27,7 @@
   --primary: #0d6dfd;
 
 
-  --secondary: #c09569;
+  --secondary: #976b3f;
   --accent: #1a55ad;
 
   --message-Q: #6426dc;
